community-agents/weather/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
pynacl>=1.5.0
pydantic>=2.6.0
